requirements.txt
pyrogram
tgcrypto
pydub
python_dotenv
git+https://github.com/openai/whisper.git
